requirements.txt
# Core dependencies
llama-index==0.11.23
llama-index-embeddings-openai==0.2.5
llama-index-vector-stores-chroma==0.2.1
chromadb==0.5.0
openai>=1.40.0

# Web scraping dependencies
beautifulsoup4==4.12.3
requests==2.32.3

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1

# Optional: For better markdown processing
markdown==3.6
markdownify==0.12.1